Added option `skipGroupSeparatorOnPaste` to numeric masks.
On the current version, when I paste "123,456.789", `,` is handled as the decimal separator and I get "132.456789".

This changes adds a `onBeforePaste` option which removes any group separators on the pasted string, fixing the problem.

I believe this should be active as default, but I'm going to be conservative and leave it off unless explicitly selected with the option `{skipGroupSeparatorOnPaste: true}`